Update dicm2nii from upstream.

+++ ext_libs/dicm2nii/nii_tool.m
@@ -78,7 +78,7 @@ function varargout = nii_tool(cmd, varargin)
 % type or dimension. The 'save' command calls this internally, so it is not
 % necessary to call this before 'save'. A useful case to call 'update' is that
 % one likes to use nii struct without saving it to a file, and 'update' will
-% make nii.hdr.dim correct.
+% make nii.hdr.dim and others correct.
 % 
 % 
 % hdr = nii_tool('hdr', filename);
@@ -226,7 +226,7 @@ function varargout = nii_tool(cmd, varargin)
 
 % History (yymmdd)
 % 150109 Write it based on Jimmy Shen's NIfTI tool (xiangrui.li@gmail.com)
-% 150202 Include renamed pigz files for Windows to trick Matlab Central
+% 150202 Include renamed pigz files for Windows
 % 150203 Fix closeFile and deleteTmpFile order
 % 150205 Add hdr.machine: needed for .img fopen
 % 150208 Add 4th input for 'save', allowing to save SPM 3D files
@@ -238,7 +238,7 @@ function varargout = nii_tool(cmd, varargin)
 % 150517 func_handle: provide a way to use gunzipOS etc from outside
 % 150617 auto detect rgb_dim 1&3 for 'load' etc using ChrisR method
 % 151025 Change subfunc img2datatype as 'update' for outside access
-% 151109 Include dd.win (exe) from WinAVR-20100110 for partial gz unzip
+% 151109 Include dd.exe from WinAVR-20100110 for partial gz unzip
 % 151205 Partial gunzip: fix fname with space & unknown pigz | dd error.
 % 151222 Take care of img for intent_code 2003/2004: anyone uses it?
 % 160110 Use matlab pref method to replace para file.
@@ -250,7 +250,6 @@ function varargout = nii_tool(cmd, varargin)
 % 161025 Make included linux pigz executible; fix "dd" for windows.
 % 161031 gunzip_mem(), nii_bytes() for hdr/ext read: read uint8 then parse;
 %        Replace hdr.machine with hdr.swap_endian.
-% 170202 check_gzip/check_dd: cd .. if m_dir is pwd, so get full path for m_dir
 % 170212 Extract decode_ext() from 'ext' cmd so call it in 'update' cmd.
 % 170215 gunzipOS: use -c > rather than copyfile for performance.
 % 170322 gzipOS: stop using background gz to avoid file not exist error.
@@ -258,6 +257,10 @@ function varargout = nii_tool(cmd, varargin)
 % 170714 'save': force to version 2 if img dim exceeds 2^15-1.
 % 170716 Add functionSignatures.json file for tab auto-completion.
 % 171031 'LocalFunc' makes eaiser to call local functions.
+% 171206 Allow file name ext other than .nii, .hdr, .img.
+% 180104 check_gzip: add /usr/local/bin to PATH for unix if needed.
+% 180119 use jsystem for better speed.
+% 180710 bug fix for cal_max/cal_min in 'update'.
 
 persistent C para; % C columns: name, length, format, value, offset
 if isempty(C)
@@ -275,7 +278,7 @@ if any(cmd=='?'), subFuncHelp(mfilename, cmd); return; end
 
 if strcmpi(cmd, 'init')
     if nargin<2, error('nii_tool(''%s'') needs second input', cmd); end
-    for i = 1:size(C,1), nii.hdr.(C{i,1}) = C{i,4}; end
+    nii.hdr = cell2struct(C(:,4), C(:,1));
     nii.img = varargin{1};
     if numel(size(nii.img))>8
         error('NIfTI img can have up to 7 dimension');
@@ -298,9 +301,7 @@ elseif strcmpi(cmd, 'save')
     % Check file name to save
     if nargin>2
         fname = varargin{2};
-        if numel(fname)<5 || ~ischar(fname)
-            error('Invalid name for NIfTI file: %s', fname);
-        end
+        if ~ischar(fname), error('Invalid name for NIfTI file: %s', fname); end
     elseif isfield(nii.hdr, 'file_name')
         fname = nii.hdr.file_name;
     else
@@ -619,10 +620,17 @@ elseif strcmpi(cmd, 'update') % old img2datatype subfunction
     nii.hdr.bitpix = para.bitpix(ind);
     nii.hdr.dim = [ndim dim];
     
-    if nii.hdr.sizeof_hdr == 348
-        nii.hdr.glmax = round(double(max(nii.img(:)))); % we may remove these
-        nii.hdr.glmin = round(double(min(nii.img(:))));
+    mx = double(max(nii.img(:)));
+    mn = double(min(nii.img(:)));
+    if nii.hdr.cal_min>mx || nii.hdr.cal_max<mn % reset wrong value
+        nii.hdr.cal_min = 0;
+        nii.hdr.cal_max = 0;
     end
+    
+    if nii.hdr.sizeof_hdr == 348
+        nii.hdr.glmax = round(mx); % we may remove these
+        nii.hdr.glmin = round(mn);
+    end    
     
     if isfield(nii, 'ext')
         try swap = nii.hdr.swap_endian; catch, swap = false; end
@@ -654,7 +662,7 @@ if niiVer == 1
     C = {
     % name              len  format     value           offset
     'sizeof_hdr'        1   'int32'     348             0
-    'data_type'         10  'char*1'    ''              4                                          
+    'data_type'         10  'char*1'    ''              4
     'db_name'           18  'char*1'    ''              14
     'extents'           1   'int32'     16384           32
     'session_error'     1   'int16'     0               36
@@ -780,13 +788,7 @@ persistent cmd; % command to gzip
 if isempty(cmd)
     cmd = check_gzip('gzip');
     if ischar(cmd)
-        if ispc
-            cmd = @(nam) sprintf('%s -nf "%s"', cmd, nam);
-            %cmd = @(nam) sprintf('start "" /B %s -nf "%s"', cmd, nam);%background
-        else
-            cmd = @(nam) sprintf('%s -nf "%s"', cmd, nam);
-            %cmd = @(nam) sprintf('%s -nf "%s" &', cmd, nam); % background
-        end
+        cmd = @(nam){cmd '-nf' nam};
     elseif islogical(cmd) && ~cmd
         fprintf(2, ['None of system pigz, gzip or Matlab gzip available. ' ...
             'Files are not compressed into gz.\n']);
@@ -798,8 +800,8 @@ if islogical(cmd)
     return;
 end
 
-[err, str] = system(cmd(fname));
-if err
+[err, str] = jsystem(cmd(fname));
+if err && ~exist([fname '.gz'], 'file')
     try
         gzip(fname); deleteFile(fname);
     catch
@@ -807,68 +809,55 @@ if err
     end
 end
 
-% Deal with pigz/gzip on path or in nii_tool folder, and matlab gzip/gunzip
+%% Deal with pigz/gzip on path or in nii_tool folder, and matlab gzip/gunzip
 function cmd = check_gzip(gz_unzip)
-m_dir = fileparts(which(mfilename));
+m_dir = fileparts(mfilename('fullpath'));
 % away from pwd, so use OS pigz if both exist. Avoid error if pwd changed later
 if strcmpi(pwd, m_dir), cd ..; clnObj = onCleanup(@() cd(m_dir)); end
+if isunix
+    pth1 = getenv('PATH');
+    if isempty(strfind(pth1, '/usr/local/bin'))
+        pth1 = [pth1 ':/usr/local/bin'];
+        setenv('PATH', pth1);
+    end
+end
 
 % first, try system pigz
-[err, ~] = system('pigz -V 2>&1');
+[err, ~] = jsystem({'pigz' '-V'});
 if ~err, cmd = 'pigz'; return; end
 
 % next, try pigz included with nii_tool
+cmd = [m_dir '/pigz'];
 if ismac % pigz for mac is not included in the package
     if strcmp(gz_unzip, 'gzip')
         fprintf(2, [' Please install pigz for fast compression: ' ...
             'http://macappstore.org/pigz/\n']);
     end
-elseif ispc % rename back pigz for Windows. Renamed to trick Matlab Central
-    try %#ok<*TRYNC>
-        fname = [m_dir '\pigz.win'];
-        if exist(fname, 'file')
-            movefile(fname, [m_dir '\pigz.exe'], 'f');
-        end
-        fname = [m_dir '\pthreadGC2.win'];
-        if exist(fname, 'file')
-            movefile(fname, [m_dir '\pthreadGC2.dll'], 'f');
-        end
-    end
-else % linux
-    [st, val] = fileattrib([m_dir '/pigz']);
-    if st && ~val.UserExecute, fileattrib([m_dir '/pigz'], '+x'); end
+elseif isunix % linux
+    [st, val] = fileattrib(cmd);
+    if st && ~val.UserExecute, fileattrib(cmd, '+x'); end
 end
 
-cmd = fullfile(m_dir, 'pigz');
-cmd = ['"' cmd '"'];
-[err, ~] = system([cmd ' -V 2>&1']);
+[err, ~] = jsystem({cmd '-V'});
 if ~err, return; end
 
 % Third, try system gzip/gunzip
-[err, ~] = system([gz_unzip ' -V 2>&1']); % gzip/gunzip on system path?
+[err, ~] = jsystem({gz_unzip '-V'}); % gzip/gunzip on system path?
 if ~err, cmd = gz_unzip; return; end
 
-% Lastly, try Matlab gzip/gunzip
-if isempty(which(gz_unzip)) || ~usejava('jvm')
-    cmd = false; % none of de/compress tools available
-    return;
-end 
-cmd = true; % use slower matlab gzip.m/gunzip.m
+% Lastly, use Matlab gzip/gunzip if java avail
+cmd = usejava('jvm');
 
 %% check dd command, return empty if not available
 function dd = check_dd
-[err, ~] = system('dd --version 2>&1');
+m_dir = fileparts(mfilename('fullpath'));
+if strcmpi(pwd, m_dir), cd ..; clnObj = onCleanup(@() cd(m_dir)); end
+[err, ~] = jsystem({'dd' '--version'});
 if ~err, dd = 'dd'; return; end % dd with linix/mac, and maybe windows
 
 if ispc % rename it as exe
-    m_dir = fileparts(which(mfilename));
-    if strcmpi(pwd, m_dir), cd ..; clnObj = onCleanup(@() cd(m_dir)); end
-    fname = [m_dir '\dd.win'];
-    if exist(fname, 'file') % first time after download
-        try movefile(fname, [m_dir '\dd.exe'], 'f'); end
-    end
-    dd = ['"' fullfile(m_dir, 'dd') '"'];
-    [err, ~] = system([dd ' --version 2>&1']);
+    dd = [m_dir '\dd'];
+    [err, ~] = jsystem({dd '--version'});
     if ~err, return; end
 end
 dd = '';
@@ -879,14 +868,14 @@ persistent cmd dd pth uid; % command to run gupzip, dd tool, and temp_path
 if isempty(cmd)
     cmd = check_gzip('gunzip'); % gzip -dc has problem in PC
     if ischar(cmd)
-        cmd = @(nam)sprintf('%s -nfdc "%s" ', cmd, nam); % f for overwrite
+        cmd = @(nam)sprintf('"%s" -nfdc "%s" ', cmd, nam); % f for overwrite
     elseif islogical(cmd) && ~cmd
         cmd = [];
         error('None of system pigz, gunzip or Matlab gunzip is available');
     end
     dd = check_dd;
     if ~isempty(dd)
-        dd = @(n,out)sprintf('| %s count=%g of="%s"', dd, ceil(n/512), out);
+        dd = @(n,out)sprintf('| "%s" count=%g of="%s"', dd, ceil(n/512), out);
     end
     
     if ispc % matlab tempdir could be slow due to cd in and out
@@ -901,6 +890,7 @@ end
 
 if islogical(cmd)
     outName = gunzip(fname, pth);
+    outName = outName{1};
     return;
 end
 
@@ -920,6 +910,7 @@ if ~isempty(dd) && nargin>1 && ~isinf(nByte) % unzip only part of data
 end
 
 [err, str] = system([cmd(fname) '> "' outName '"']);
+% [err, str] = jsystem({'pigz' '-nfdc' fname '>' outName});
 if err
     try
     	outName = gunzip(fname, pth);
@@ -1000,7 +991,7 @@ for i = 1:numel(ext)
         fclose(fid1);
         deleteMat = onCleanup(@() deleteFile(tmp)); % delete temp file after done
         ext(i).edata_decoded = load(tmp); % load into struct
-    elseif ext(i).ecode == 6 % plain text
+    elseif any(ext(i).ecode == [4 6 32]) % 4 AFNI, 6 plain text, 32 CIfTI
         str = char(ext(i).edata(:)');
         if isempty(strfind(str, 'dicm2nii.m'))
             ext(i).edata_decoded = deblank(str);
@@ -1015,9 +1006,14 @@ for i = 1:numel(ext)
                 if isempty(a), continue; end
                 try
                     eval(['ss.' a]); % put all into struct
-                catch me
-                    fprintf(2, '%s\n', me.message);
-                    fprintf(2, 'Unrecognized text: %s\n', a);
+                catch
+                    try
+                        a = regexp(a, '(.*?)\s*=\s*(.*?);', 'tokens', 'once');
+                        ss.(a{1}) = a{2};
+                    catch me
+                        fprintf(2, '%s\n', me.message);
+                        fprintf(2, 'Unrecognized text: %s\n', a);
+                    end
                 end
             end
             flds = fieldnames(ss); % make all vector column
@@ -1027,8 +1023,6 @@ for i = 1:numel(ext)
             end
             ext(i).edata_decoded = ss;
         end
-    elseif ext(i).ecode == 4 % AFNI
-        ext(i).edata_decoded = deblank(char(ext(i).edata(:))');
     elseif ext(i).ecode == 2 % dicom
         tmp = [tempname '.dcm'];
         fid1 = fopen(tmp, 'W');
@@ -1051,7 +1045,6 @@ dim = hdr.dim(2:8);
 dim(hdr.dim(1)+1:7) = 1; % avoid some error in file
 dim(dim<1) = 1;
 valpix = para.valpix(ind);
-n = prod(dim); % num of voxels
 
 fname = nii_name(hdr.file_name, '.img'); % in case of .hdr/.img pair
 fid = fopen(fname);
@@ -1063,6 +1056,13 @@ if isequal(sig, [31 139]) % .gz
     fid = fopen(fname);
 end
 
+% if ~exist('cln', 'var') && valpix==1 && ~hdr.swap_endian
+%     m = memmapfile(fname, 'Offset', hdr.vox_offset, ...
+%         'Format', {para.format{ind}, dim, 'img'});
+%     nii = m.Data;
+%     return;
+% end
+
 if hdr.swap_endian % switch between LE and BE
     [~, ~, ed] = fopen(fid); % default endian: almost always ieee-le
     fclose(fid);
@@ -1073,7 +1073,7 @@ if hdr.swap_endian % switch between LE and BE
 end
 
 fseek(fid, hdr.vox_offset, 'bof');
-img = fread(fid, n*valpix, ['*' para.format{ind}]); % * to keep original class
+img = fread(fid, prod(dim)*valpix, ['*' para.format{ind}]); % * to keep original class
 fclose(fid);
 
 if any(hdr.datatype == [128 511 2304]) % RGB or RGBA
@@ -1103,17 +1103,13 @@ end
 
 %% Return requested fname with ext, useful for .hdr and .img files
 function fname = nii_name(fname, ext)
-[~, f, e] = fileparts(fname);
-n = numel(fname);
-if strcmpi(e, '.gz')
-    n = n - 3; % 3 is numel('.gz')
-    [~, ~, e] = fileparts(f); % .nii/.hdr/.img
+if strcmpi(ext, '.img')
+    i = regexpi(fname, '.hdr(.gz)?$');
+    if ~isempty(i), fname(i(end)+(0:3)) = ext; end
+elseif strcmpi(ext, '.hdr') 
+    i = regexpi(fname, '.img(.gz)?$');
+    if ~isempty(i), fname(i(end)+(0:3)) = ext; end
 end
-if strcmpi(e, '.nii') || strcmpi(e, ext), return; end
-if ~strcmpi(e, '.hdr') && ~strcmpi(e, '.img')
-    error(['Invalid NIfTI file name: ' fname]); 
-end
-fname(n+(-3:0)) = ext; % if not return or error, change it
 
 %% Read NIfTI file as bytes, gunzip if needed, but ignore endian
 function [b, fname] = nii_bytes(fname, nByte)
@@ -1170,11 +1166,14 @@ function bytes = gunzip_mem(gz_bytes, fname, nByte)
 bytes = [];
 try
     bais = java.io.ByteArrayInputStream(gz_bytes);
-    gzis = java.util.zip.GZIPInputStream(bais);
+    try, gzis = java.util.zip.GZIPInputStream(bais); %#ok<*NOCOM>
+    catch, try, gzis = java.util.zip.InflaterInputStream(bais); catch me; end
+    end
     buff = java.io.ByteArrayOutputStream;
-    try org.apache.commons.io.IOUtils.copy(gzis, buff); end
+    try org.apache.commons.io.IOUtils.copy(gzis, buff); catch me; end
     gzis.close;
-    bytes = typecast(buff.toByteArray, 'uint8'); % faster than uint8()
+    bytes = typecast(buff.toByteArray, 'uint8');
+    if isempty(bytes), error(me.message); end
 catch
     if nargin<3 || isempty(nByte), nByte = inf; end
     if nargin<2 || isempty(fname)
@@ -1186,7 +1185,7 @@ catch
         fclose(fid);
     end
     
-    try
+    try %#ok<*TRYNC>
         fname = gunzipOS(fname, nByte);
         fid = fopen(fname);
         bytes = fread(fid, nByte, '*uint8');
@@ -1199,5 +1198,21 @@ end
 function deleteFile(fname)
 if ispc, system(['start "" /B del "' fname '"']);
 else, system(['rm "' fname '" &']);
+end
+
+%% faster than system: based on https://github.com/avivrosenberg/matlab-jsystem
+function [err, out] = jsystem(cmd)
+% cmd is cell str, no quotation marks needed for file names with space.
+try
+    pb = java.lang.ProcessBuilder(cmd);
+    pb.redirectErrorStream(true); % ErrorStream to InputStream
+    process = pb.start();
+    scanner = java.util.Scanner(process.getInputStream).useDelimiter('\A');
+    if scanner.hasNext(), out = char(scanner.next()); else, out = ''; end
+    err = process.exitValue; % err = process.waitFor() may hang
+    if err, error('java.lang.ProcessBuilder error'); end
+catch % fallback to system() if java fails like for Octave
+    cmd = regexprep(cmd, '.+? .+', '"$0"'); % double quotes if with middle space
+    [err, out] = system(sprintf('%s ', cmd{:}));
 end
 %%
